+++ opteryx/engine/query_statistics.py
@@ -32,6 +32,8 @@ class QueryStatistics:
         self.partitions_read: int = 0
         self.time_scanning_partitions: int = 0
 
+        self.segments_scanned: int = 0
+
         self.time_data_read: int = 0
 
         self.cache_hits: int = 0
@@ -72,6 +74,7 @@ class QueryStatistics:
             "partitions_found": self.partitions_found,
             "partitions_scanned": self.partitions_scanned,
             "partitions_read": self.partitions_read,
+            "segments_scanned": self.segments_scanned,
             "cache_hits": self.cache_hits,
             "cache_misses": self.cache_misses,
         }